master: refactor signals handling
